update mongodb to ts
